msgraph: fully migrate to new structure
