+++ packages/common/src/Adaptor.js
@@ -13,21 +13,9 @@ export * as beta from './beta';
 export * as http from './http.deprecated';
 export * as dateFns from './dateFns';
 
-const schemaCache = {};
+export { each, fn, sourceValue } from './operations'
 
-// operation creates a new operation factory
-// It accepts a function which takes user arguments AND state
-// It returns a function which acccepts user args
-// Which in turn returns a function which accepts state
-// and then triggers the original function
-// It's complex BUT the end user code should be way way simpler
-export const operation = (fn) => {
-  return (...args) => {
-    return (state) => {
-      return fn(state, ...args)
-    }
-  }
-}
+const schemaCache = {};
 
 
 /**
