+++ packages/msgraph/test/Adaptor.deprecated
