--- packages/msgraph/test/mock.js
@@ -1,4 +1,0 @@
-const initMock = (endpoints) => {
-  // here we basically need to override the request function
-  // Ho do we do this?
-}
